tests/wpt/tests/fedcm/fedcm-token-endpoint-rejects-redirects.https.html

<!DOCTYPE html>
<title>Federated Credential Management API token endpoint does not follow redirects.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
import {request_options_with_mediation_required,
        fedcm_test,
        select_manifest,
        fedcm_error_dialog_dismiss,
        fedcm_get_and_select_first_account} from './support/fedcm-helper.sub.js';

fedcm_test(async t => {
  let test_options = request_options_with_mediation_required("manifest_redirect_token.json");
  await select_manifest(t, test_options);

  try {
    const cred = fedcm_get_and_select_first_account(t, test_options);
    fedcm_error_dialog_dismiss(t);
    await cred;
    assert_unreached("An IdentityCredentialError exception should be thrown.");
  } catch (e) {
    assert_true(e instanceof DOMException);
    assert_equals(e.name, "IdentityCredentialError");
    // 308 should not produce a valid error code
    assert_equals(e.error, "");
  }
}, 'Test that token endpoint does not follow redirects');
  </script>
